Update plugins to the latest version
This involves a big switch in the way LSPs are setup. Previously, the
"handlers" pattern was used with mason-lspconfig, but now you basically
just want to call "vim.lsp.config(...)" with the config. Mason also deprecated
get_install_directory, which you can basically replace with the $MASON
environment variable. This removes the need for mason-lspconfig all
together. It is likely lspconfig itself will go down the same road in a
future nvim release, but it stays for now as it provides default configs
via runtimepath/lsp/...