Fix Filetypes not being detected
On the 1st buffer load, this would happen starting with neovim v0.11.2.
This would cause Treesitter not to load, meaning missing syntax
highlighting among other things. The reason is that LSPConfig should now
happen BufReadPre event to set the filetype in time.

This also takes the opportunity to put the "Lazy File" events in one
place, so that if something like this ever happens again it's a little
bit easier to deal with.